because there isn't really that much code in each of the controllers, reduce nesting and bring them to their own level. api controller just has functions that can be used rather than having to instantiate it
